+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Added already exists error.
+- Added resource already exists error.
 
 ## [4.5.0] - 2021-03-26